dockerFiles/requirements.txt
flask
requests
flask_cors
gitpython